+++ r4300/cop1_l.c
@@ -34,7 +34,7 @@ void CVT_S_L()
 {
    if (check_cop1_unusable()) return;
    set_rounding();
-   *reg_cop1_simple[cffd] = *((long long*)reg_cop1_double[cffs]);
+   *reg_cop1_simple[cffd] = *((s64*)reg_cop1_double[cffs]);
    PC++;
 }
 
@@ -42,6 +42,6 @@ void CVT_D_L()
 {
    if (check_cop1_unusable()) return;
    set_rounding();
-   *reg_cop1_double[cffd] = *((long long*)reg_cop1_double[cffs]);
+   *reg_cop1_double[cffd] = *((s64*)reg_cop1_double[cffs]);
    PC++;
 }
