+++ r4300/cop1_w.c
@@ -35,7 +35,7 @@ void CVT_S_W()
 {  
    if (check_cop1_unusable()) return;
    set_rounding();
-   *reg_cop1_simple[cffd] = *((long*)reg_cop1_simple[cffs]);
+   *reg_cop1_simple[cffd] = *((s32*)reg_cop1_simple[cffs]);
    PC++;
 }
 
@@ -43,6 +43,6 @@ void CVT_D_W()
 {
    if (check_cop1_unusable()) return;
    set_rounding();
-   *reg_cop1_double[cffd] = *((long*)reg_cop1_simple[cffs]);
+   *reg_cop1_double[cffd] = *((s32*)reg_cop1_simple[cffs]);
    PC++;
 }
